:sparkles: Adds json exports
